XMLHttpRequest/send-no-response-event-loadend.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The send() method: Fire a progress event named loadend (no response entity body)</title>
</head>

<body>
    <div id="log"></div>
    
    <script type="text/javascript">
        var test = async_test();

        test.step(function () 
        {
            var xhr = new XMLHttpRequest();

            xhr.onreadystatechange = function() 
            {
                test.step(function() 
                {
                    if (xhr.readyState == 4) 
                    {
                        assert_equals(xhr.response, "");
                    }
                });
            };

            xhr.onloadend = function(e) 
            {
                test.step(function() 
                {
                    assert_true(e instanceof ProgressEvent);
                    assert_equals(e.type, "loadend");
                    test.step(function() { test.done(); });
                });
            };

            xhr.open("POST", "./support/echo.php", true);
            xhr.send();
        });
    </script>
</body>
</html>
